Give every ML experiment a baseline, and report the two that fail it
The predictive-maintenance experiment had no baseline, and that is what let it
pass. Ranking the maintenance queue by asset age alone scores ROC-AUC 0.620;
the model scored 0.598. It had been losing to a single variable while
reporting a 1.35x lift that turns out to be noise -- top-5% lift moves between
1.03 and 1.35 across capacity settings with no trend. Capacity is now heavily
regularised, the age-only baseline is computed and logged on every run, and
the experiment prints why it does not beat it: breakdown probability in the
generator depends on a latent per-asset propensity that no column observes
except through age, and 61% of assets have exactly one work order, so the
history features are empty for most rows. Fixing that means adding condition
monitoring to the generator, not tuning this file.

The stock-out model was scoring ROC-AUC 0.998, which the project's own stated
rule should have caught: the target is days_of_cover below a threshold,
days_of_cover is stock over usage, and both were features. Excluding
days_of_cover -- which the code did, with a comment explaining why -- excluded
nothing, because the model just divides one component by the other. Re-pointed
at the next snapshot of the same tank, where the target is measured after
every feature, it scores 0.510. The median gap between snapshots is 84 days
and a tank level today says nothing about a level a quarter later, so at this
cadence the model does not exist. The leaked version was the only learnable
one, which is why the leak was worth finding. Feeding it from the tank
telemetry stream instead of quarterly snapshots is the fix.

Also added the condition features the maintenance model should have had
(time since last work order, time since last breakdown, work in the last 90
days, the site's own reliability). They contribute nothing here for the reason
above, and are kept because they are correct and the reason they are flat is
documented.

late_delivery had been failing outright: DuckDB returns nullable Int64 for
drop_count and ColumnTransformer refuses to put pd.NA in a numpy array, so it
raised at fit with a traceback pointing at the pipeline rather than at the
column. as_model_matrix in ml/common.py now coerces nullable numerics once,
for every experiment. It scores 0.564 with 1.76x lift in the top decile --
marginal, but a reviewer working that decile finds 33% late against an 18%
base rate.

A predictive-maintenance rolling window was also dropping rows silently:
work orders with a null raise timestamp cannot be placed in a time series at
all, and the positional assignment either raised on the length mismatch or
would have attached values to the wrong rows. They are now dropped explicitly
with a count, and each group's window is rebuilt on its own index.

README now lists all six models with the two failures stated as failures.

+++ databricks/notebooks/00_generate_bronze_at_scale.py
@@ -51,7 +51,7 @@ for key, value in [
 ]:
     try:
         spark.conf.set(key, value)
-    except Exception:                                        # noqa: BLE001
+    except Exception:
         print(f"note: {key} is managed by the runtime, leaving as-is")
 
 SCALES = {
@@ -434,7 +434,7 @@ HOUR_WEIGHTS = [0.22, 0.14, 0.10, 0.10, 0.18, 0.45, 0.95, 1.55, 1.70, 1.25,
                 0.85, 0.62, 0.45, 0.30]
 hour_slots = []
 for h, w in enumerate(HOUR_WEIGHTS):
-    hour_slots += [h] * int(round(w * 100))
+    hour_slots += [h] * round(w * 100)
 hour_arr = F.array(*[F.lit(h) for h in hour_slots])
 n_hour_slots = len(hour_slots)
 
@@ -726,7 +726,7 @@ rows = []
 for t in sorted(tables):
     try:
         rows.append((t, spark.table(f"{CATALOG}.bronze.{t}").count()))
-    except Exception as exc:                        # noqa: BLE001
+    except Exception as exc:
         rows.append((t, -1))
         print(f"could not count {t}: {exc}")
